+++ tests/mocks/mock_country.py
@@ -1,0 +1,8 @@
+class MockCountry:
+  def __init__(self, *args, **kwargs):
+    self.iso = 'abc'
+    self.date = '2021-01-01' 
+
+  @staticmethod
+  def set_data(*args, **kwargs):
+    pass